Create analysis submodule
