+++ tests/OTAnalytics/plugin_ui/test_cli.py
@@ -9,7 +9,10 @@ import pytest
 from OTAnalytics.adapter_intersect.intersect import (
     ShapelyIntersectImplementationAdapter,
 )
-from OTAnalytics.application.analysis import RunIntersect, RunSceneEventDetection
+from OTAnalytics.application.analysis.analysis import (
+    RunIntersect,
+    RunSceneEventDetection,
+)
 from OTAnalytics.application.datastore import (
     EventListParser,
     SectionParser,
